added delete functionality
