+++ friends/src/List/FriendList.js
@@ -7,7 +7,14 @@ const FriendList = props => {
   return (
 
       <div>
-            {props.friendsData.map(friend => <p key={friend.id}>{friend.name} {friend.age} {friend.email}</p>)}
+            {props.friendsData.map(friend => {
+		    return(	
+		    <div className="list-container" key={friend.id}>		    
+		    <p>{friend.name} {friend.age} {friend.email}</p>
+		    <button className="delete-btn" onClick={()=>props.deleteFriend(friend.id)}>Delete</button>
+		    </div>		    
+		    ); 
+	    })}
       </div>
   );
 }
